Move files in the appointment book GWT original project to the "apptbookgwt" package so that we can have unique package names in the original projects that will not confuse the aggregate javadoc build.
