+++ gwt-parent/gwt-originals-parent/apptbook-gwt/src/it/java/edu/pdx/cs410J/apptbookgwt/client/AppointmentBookGwtIT.java
@@ -1,4 +1,4 @@
-package edu.pdx.cs410J.original.client;
+package edu.pdx.cs410J.apptbookgwt.client;
 
 import com.google.gwt.dom.client.Document;
 import com.google.gwt.dom.client.NativeEvent;
@@ -16,7 +16,7 @@ import org.junit.Test;
 public class AppointmentBookGwtIT extends GWTTestCase {
   @Override
   public String getModuleName() {
-    return "edu.pdx.cs410J.original.AppointmentBookIntegrationTests";
+    return "edu.pdx.cs410J.apptbookgwt.AppointmentBookIntegrationTests";
   }
 
   @Test
