+++ gwt-parent/gwt-originals-parent/apptbook-gwt/src/it/java/edu/pdx/cs410J/apptbookgwt/client/AppointmentBookGwtTestSuite.java
@@ -1,4 +1,4 @@
-package edu.pdx.cs410J.original.client;
+package edu.pdx.cs410J.apptbookgwt.client;
 
 import com.google.gwt.junit.tools.GWTTestSuite;
 import junit.framework.Test;
