+++ gwt-parent/gwt-originals-parent/apptbook-gwt/src/it/resources/edu/pdx/cs410J/apptbookgwt/AppointmentBookIntegrationTests.gwt.xml
@@ -8,7 +8,7 @@
 
   -->
 
-  <inherits name='edu.pdx.cs410J.original.AppointmentBookGwt'/>
+  <inherits name='edu.pdx.cs410J.apptbookgwt.AppointmentBookGwt'/>
 
   <source path="client">
     <include name="**/*.java"/>
@@ -22,6 +22,6 @@
   </source>
 
   <!-- Need to define the servlet in .gwt.xml in order to use service in a GwtTest -->
-  <servlet path="/ping" class="edu.pdx.cs410J.original.server.PingServiceImpl"/>
+  <servlet path="/ping" class="edu.pdx.cs410J.apptbookgwt.server.PingServiceImpl"/>
 
 </module>
