+++ gwt-parent/gwt-originals-parent/apptbook-gwt/src/main/java/edu/pdx/cs410J/apptbookgwt/client/Appointment.java
@@ -1,4 +1,4 @@
-package edu.pdx.cs410J.original.client;
+package edu.pdx.cs410J.apptbookgwt.client;
 
 import edu.pdx.cs410J.AbstractAppointment;
